requirements.txt
pandas
ffats==1.3.11
tqdm
requests
numpy
scikit-learn==1.4.2
torch==2.2.1
torchaudio
fastapi
uvicorn
python-multipart
python-dotenv
aiogram
httpx
